Remove uses of %l
This change removes uses of %l from the EC side of the EC codebase.
This is done because the semantics of %l within printf have changed,
and there are concerns that new calls to printf will be cherry-picked
into old firmware branches without the printf changes. So, in
preparation for disallowing %l in master, remove occurrences of %l.

This change was done by manually fixing up anything found under the EC
directory with the following regex: %[0-9*.-]*l[^l]

Remember that anything on the host machine is fine as-is, since the host
printf never changed.

BUG=chromium:984041
TEST=make -j buildall
BRANCH=None

+++ core/minute-ia/task.c
@@ -253,7 +253,9 @@ uint32_t switch_handler(int desched, task_id_t resched)
 		return 0;
 
 	if (IS_ENABLED(ISH_DEBUG))
-		CPRINTF("[%ld -> %ld]\n", current - tasks, next - tasks);
+		CPRINTF("[%lld -> %lld]\n",
+			(long long)(current - tasks),
+			(long long)(next - tasks));
 
 	/* Switch to new task */
 	if (IS_ENABLED(CONFIG_TASK_PROFILING))
